First prototype (at least its working)
Version 0.0.0.0.0.8
this is pretty much the entire concept of the game, obviously it ain't pretty right now, I will make it prettier once I receive the animations and all that, for now I will just focus on upgrading the gameplay, adding game over screen and all that, will work on the display as well. keep in mind that this is the Player vs Player mode. I will be working on the Player vs AI a bit later.