Beta version 0.0.1
This is a test of the beta. What's different here is the way the player restarts after GameOver and we implemented a shoot to start screen so that the game doesn't start as soon as you launch. Still missing the crouching visual but nothing I can do about that since I have yet to receive the characters.